Add missinc cmath include and use std::pow instead of std::powf

+++ Source/arch/Range.cpp
@@ -1,4 +1,5 @@
 #include "Range.h"
+#include <cmath>
 
 namespace makeRange
 {
@@ -50,11 +51,11 @@ namespace makeRange
 			start, end,
 			[b = f1k, bInv = 1.f / f1k, r = start - end](float, float max, float x)
 			{
-				return max + r * std::powf(1.f - std::powf(x, b), bInv);
+				return max + r * std::pow(1.f - std::pow(x, b), bInv);
 			},
 			[b = k1f, bInv = 1.f / k1f, rInv = 1.f / (end - start)](float min, float, float x)
 			{
-				return 1.f - std::powf(1.f - std::powf((x - min) * rInv, b), bInv);
+				return 1.f - std::pow(1.f - std::pow((x - min) * rInv, b), bInv);
 			},
 			[](float min, float max, float x)
 			{
